Subject: [PATCH 4/4] UsbmonLogParser: add setup packet to correct transactions
Requests to CONTROL endpoints (ie, to enpoint 0) are logged differently
in VMX logs and in usbmon logs. VMX adds the setup packet to both the
Down part and the Up part of a transaction. usbmon only adds the setup
packet to the Down part of those transactions.

UsbmonLogParser currently tries to add the setup packet to some
transactions but doesn't only add it to the Up part of request to
endpoint 0. A simple rewrite of the way setup packets are added to
transactions in usbmon logs seems to be doing the right thing.

Signed-off-by: Paul Bolle <pebolle@tiscali.nl>



git-svn-id: file:///tmp/tmp.CAC7Htwj4j/svn/trunk@24 31ffde5e-7fae-4e65-b3f5-864f19086f7f